requirements.txt
attrs==21.2.0
autopep8==1.6.0
click==8.0.3
Flask==2.0.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1
packaging==21.2
pluggy==1.0.0
psycopg2==2.9.2
py==1.11.0
pycodestyle==2.8.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.25
SQLAlchemy-Utils==0.37.9
toml==0.10.2
Werkzeug==2.0.2
